showcase: fix player to work well on HiDPI, have working fullsize etc.
Similarly to mouseinteraction, the new EmscriptenApplication has a
different window and framebuffer size on HiDPI and the code didn't
account for that. Moreover, the full size switcher was broken due to an
unreferenced function on the JS side *and* viewport event not being
triggered on canvas resize. I wonder who tested this :(

Also discovered a few issues with object getting randomly rotated after
clicking the canvas while the browser was in background. Turns out that
was due to unexpected order of events coming from the browser (move
event with buttons pressed being fired *before* mouse press event).
Worked around that.