showcase: use up-to-date HTML markup for all.
This change was done ages ago for the examples and other applications,
but here it stayed with the original IDs instead of CSS classes. Which
means that, for example, the latest magnum-player binary doesn't have
the "Full size" button working because it applies a CSS class and not an
ID.

Also added new aspect ratio and size variants to the CSS file to make it
in sync with Magnum's own WebApplication.css.