showcase: update player to latest with additional plugins compiled in.
And list what all file formats are actually supported. It's now a 3.7 MB
binary, but ... given current web "standards" it might be completely
fine.

The biggest offenders there are OpenEXR (~0.5 MB, because it includes a
ton of compression algorithms and, well, does *a lot*), and Basis (~0.5
MB as well, because goddamit I have no idea what's going on with that
codebase, come on...).